functions/parse-object-metadata/requirements.txt
pytz>=2018.7
iso8601>=0.1.12
google-cloud-storage>=1.15.0
google-cloud-pubsub>=0.40.0
google-cloud-logging>=3.0.0
trellisdata>=0.1.17
pyyaml>=6.0.1
